Refactor: Reorder protocol messages in better groups
Feat: Add 4-digits decimal precision option for protocol messages

Refactor: Make speed ki float instead of uint16_t

Refactor: Break down debug prints for encoder information into multiple functions for easier selection of what to print

Feat: Add ping messages to the serial protocol

+++ Core/pid/include/pid/controllers/speed_pid.h
@@ -44,7 +44,7 @@ void set_base_speed_kp(const uint16_t kp);
  * @brief Set the integral gain (Ki) for the speed PID controller.
  * @param ki The new integral gain value.
  */
-void set_base_speed_ki(const uint16_t ki);
+void set_base_speed_ki(const float ki);
 
 /**
  * @brief Set the derivative gain (Kd) for the speed PID controller.
